+++ WebAPI/Models/SalesRecord.cs
@@ -4,7 +4,7 @@ namespace WebAPI.Models
 {
     public class SalesRecord
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public DateTime Date { get; set; }
         public Double Amount { get; set; }
         public SaleStatus Status { get; set; }
@@ -15,7 +15,7 @@ namespace WebAPI.Models
         {
 
         }
-                public SalesRecord(long id, DateTime date, double amount, SaleStatus status, Seller seller)
+                public SalesRecord(int id, DateTime date, double amount, SaleStatus status, Seller seller)
         {
             Id = id;
             Date = date;
